wip(context): created ContextManager
