+++ package.json
@@ -57,6 +57,7 @@
   },
   "main": "index.js",
   "dependencies": {
+    "@sasjs/utils": "^1.3.0",
     "es6-promise": "^4.2.8",
     "form-data": "^3.0.0",
     "isomorphic-fetch": "^2.2.1"
